+++ packages/common/src/hooks/index.ts
@@ -21,6 +21,7 @@ export * from './useSavedCollections'
 export * from './chats'
 export * from './useGeneratePlaylistArtwork'
 export * from './useUSDCBalance'
+export * from './useAudioBalance'
 export * from './purchaseContent'
 export * from './useAccessAndRemixSettings'
 export * from './useInterval'
