+++ packages/mobile/src/components/usdc-balance-pill/UsdcBalancePill.tsx
@@ -31,7 +31,7 @@ const useStyles = makeStyles(({ spacing, palette }) => ({
   }
 }))
 
-export const USDCPill = () => {
+export const UsdcBalancePill = () => {
   const styles = useStyles()
   const { data: usdcBalance, balanceStatus: usdcBalanceStatus } =
     useUSDCBalance({ isPolling: false })
