+++ packages/mobile/src/screens/app-drawer-screen/left-nav-drawer/LeftNavDrawer.tsx
@@ -20,8 +20,8 @@ import IconMessage from 'app/assets/images/iconMessage.svg'
 import IconSettings from 'app/assets/images/iconSettings.svg'
 import IconUpload from 'app/assets/images/iconUpload.svg'
 import IconUser from 'app/assets/images/iconUser.svg'
-import { AudioPill } from 'app/components/audio-pill/AudioPill'
-import { USDCPill } from 'app/components/usdc-pill/USDCPill'
+import { AudioBalancePill } from 'app/components/audio-balance-pill/AudioBalancePill'
+import { UsdcBalancePill } from 'app/components/usdc-balance-pill/UsdcBalancePill'
 import { useFeatureFlag, useRemoteVar } from 'app/hooks/useRemoteConfig'
 import { make, track } from 'app/services/analytics'
 import { spacing } from 'app/styles/spacing'
@@ -95,7 +95,7 @@ const WrappedLeftNavDrawer = () => {
         to='PayAndEarnScreen'
         params={null}
       >
-        <USDCPill />
+        <UsdcBalancePill />
       </LeftNavLink>
       <LeftNavLink
         icon={IconCrown}
@@ -104,7 +104,7 @@ const WrappedLeftNavDrawer = () => {
         params={null}
         showNotificationBubble={hasClaimableRewards}
       >
-        <AudioPill />
+        <AudioBalancePill />
       </LeftNavLink>
       <LeftNavLink
         icon={IconUpload}
