+++ packages/web/src/components/audio-balance-pill/AudioBalancePill.tsx
@@ -1,12 +1,11 @@
-import { cloneElement, useMemo } from 'react'
+import { cloneElement } from 'react'
 
 import {
-  BNWei,
   accountSelectors,
   formatWei,
   isNullOrUndefined,
   useSelectTierInfo,
-  walletSelectors
+  useTotalBalanceWithFallback
 } from '@audius/common'
 import BN from 'bn.js'
 import cn from 'classnames'
@@ -16,39 +15,15 @@ import IconNoTierBadge from 'assets/img/tokenBadgeNoTier.png'
 import Skeleton from 'components/skeleton/Skeleton'
 import { audioTierMapPng } from 'components/user-badges/UserBadges'
 
-import styles from './AudioPill.module.css'
+import styles from './AudioBalancePill.module.css'
 
 const { getAccountUser } = accountSelectors
-const { getAccountTotalBalance } = walletSelectors
 
 type AudioPillProps = {
   className?: string
 }
 
-/**
- * Pulls balances from account and wallet selectors. Will prefer the wallet
- * balance once it has loaded. Otherwise, will return the account balance if
- * available. Falls back to 0 if neither wallet or account balance are available.
- */
-const useTotalBalanceWithFallback = () => {
-  const account = useSelector(getAccountUser)
-  const walletTotalBalance = useSelector(getAccountTotalBalance)
-
-  return useMemo(() => {
-    if (!isNullOrUndefined(walletTotalBalance)) {
-      return walletTotalBalance
-    } else if (
-      !isNullOrUndefined(account) &&
-      !isNullOrUndefined(account.total_balance)
-    ) {
-      return new BN(account.total_balance) as BNWei
-    }
-
-    return null
-  }, [account, walletTotalBalance])
-}
-
-export const AudioPill = ({ className }: AudioPillProps) => {
+export const AudioBalancePill = ({ className }: AudioPillProps) => {
   const account = useSelector(getAccountUser)
   const totalBalance = useTotalBalanceWithFallback()
   const positiveTotalBalance = !isNullOrUndefined(totalBalance)
