+++ packages/web/src/components/nav/desktop/NavPopupMenu.module.css
@@ -25,7 +25,7 @@
   position: relative;
 }
 
-.icon>svg path {
+.icon > svg path {
   fill: var(--neutral-light-4);
 }
 
@@ -68,7 +68,6 @@
   width: var(--unit-5);
   height: var(--unit-5) !important;
 }
-
 
 .container {
   display: flex;
